Integration models

Add integration model to a psdesign object

Usage

add_integration(psdesign, integration)

Arguments

psdesign

A psdesign object

integration

An integration object

Details

This is a list of the available integration models. The fundamental problem in surrogate evaluation is that there are unobserved values of the counterfactual surrogate responses S(1). In the estimated maximum likelihood framework, for subjects missing the S(1) values, we use an auxiliary pre-treatment variable or set of variables W that is observed for every subject to estimate the distribution of S(1) | W. Typically, this W is a BIP. Then for each missing S(1), we integrate likelihood contributions over each non-missing S(1) given their value of W, and average over the contributions.

integrate_parametric This is a parametric integration model that fits a linear model for the mean of S(1) | W and assumes a Gaussian distribution.

integrate_bivnorm This is another parametric integration model that assumes that S(1) and W are jointly normally distributed. The user must specify their mean, variances and correlation.

integrate_nonparametric This is a non-parametric integration model that is only valid for categorical S(1) and W. It uses the observed proportions to estimate the joint distribution of S(1), W.

integrate_semiparametric This is a semi-parametric model that uses the semi-parametric location scale model of Heagerty and Pepe (1999). Models are specified for the location of S(1) | W and the scale of S(1) | W. Then integrations are drawn from the empirical distribution of the residuals from that model, which are then transformed to the appropriate location and scale.

Examples

test <- psdesign(generate_example_data(n = 100), Z = Z, Y = Y.obs, S = S.obs, BIP = BIP)
add_integration(test, integrate_parametric(S.1 ~ BIP))
#> Augmented data frame:  100  obs. by  6  variables. 
#>   Z Y  S.1    S.0 cdfweights    BIP
#> 1 0 1   NA -0.739          1 -0.891
#> 2 0 0   NA -0.113          1 -0.117
#> 3 0 0   NA -0.868          1 -0.942
#> 4 1 0 1.97     NA          1  1.116
#> 5 1 0 1.45     NA          1  0.470
#> 6 1 0 2.02     NA          1  0.861
#> 
#> Empirical TE:  0.267 
#> 
#> Mapped variables: 
#> 	 Z  ->  Z 
#> 	 Y  ->  Y.obs 
#> 	 S  ->  S.obs 
#> 	 BIP  ->  BIP 
#> 
#> Integration models: 
#> 	 integration model for  S.1 :
#> 		 integrate_parametric(formula = S.1 ~ BIP )
#> 
#> Risk models: 
#> 	None present, see ?add_riskmodel for information on risk models.
#> 	No estimates present, see ?ps_estimate.
#> 	No bootstraps present, see ?ps_bootstrap.
test + integrate_parametric(S.1 ~ BIP)  # same as above
#> Augmented data frame:  100  obs. by  6  variables. 
#>   Z Y  S.1    S.0 cdfweights    BIP
#> 1 0 1   NA -0.739          1 -0.891
#> 2 0 0   NA -0.113          1 -0.117
#> 3 0 0   NA -0.868          1 -0.942
#> 4 1 0 1.97     NA          1  1.116
#> 5 1 0 1.45     NA          1  0.470
#> 6 1 0 2.02     NA          1  0.861
#> 
#> Empirical TE:  0.267 
#> 
#> Mapped variables: 
#> 	 Z  ->  Z 
#> 	 Y  ->  Y.obs 
#> 	 S  ->  S.obs 
#> 	 BIP  ->  BIP 
#> 
#> Integration models: 
#> 	 integration model for  S.1 :
#> 		 integrate_parametric(formula = S.1 ~ BIP )
#> 
#> Risk models: 
#> 	None present, see ?add_riskmodel for information on risk models.
#> 	No estimates present, see ?ps_estimate.
#> 	No bootstraps present, see ?ps_bootstrap.