Add risk model to a psdesign object

Add risk model to a psdesign object

Usage

add_riskmodel(psdesign, riskmodel)

Arguments

psdesign

A psdesign object

riskmodel

A risk model object, from the list above

Details

The risk model component specifies the likelihood for the data. This involves specifying the distribution of the outcome variable, whether it is binary or time-to-event, and specifying how the surrogate S(1) and the treatment Z interact and affect the outcome. We use the formula notation to be consistent with other regression type models in R. Below is a list of available risk models.

risk_binary This is a generic risk model for binary outcomes. The user can specify the formula, and link function using either risk.logit for the logistic link, or risk.probit for the probit link. Custom link functions may also be specified, which take a single numeric vector argument, and returns a vector of corresponding probabilities.

risk_weibull This is a parameterization of the Weibull model for time-to-event outcomes that is consistent with that of rweibull. The user specifies the formula for the linear predictor of the scale parameter.

risk_exponential This is a simple exponential model for a time-to-event outcome.

risk_poisson This is a Poisson model for count outcomes. It allows for offsets in the formula.

risk_continuous This is a Gaussian model for continuous outcomes. It assumes that larger values of the outcome are harmful (e.g. blood pressure)

Examples

test <- psdesign(generate_example_data(n = 100), Z = Z, Y = Y.obs, S = S.obs, BIP = BIP) +
     integrate_parametric(S.1 ~ BIP)
add_riskmodel(test, risk_binary())
#> Augmented data frame:  100  obs. by  6  variables. 
#>   Z Y    S.1    S.0 cdfweights      BIP
#> 1 0 0     NA -0.542          1 -0.51164
#> 2 1 0  1.095     NA          1 -0.00197
#> 3 0 0     NA  1.751          1  1.78313
#> 4 0 1     NA  1.062          1  1.03138
#> 5 0 0     NA -0.501          1 -0.58070
#> 6 1 0 -0.422     NA          1 -1.59308
#> 
#> Empirical TE:  0.744 
#> 
#> Mapped variables: 
#> 	 Z  ->  Z 
#> 	 Y  ->  Y.obs 
#> 	 S  ->  S.obs 
#> 	 BIP  ->  BIP 
#> 
#> Integration models: 
#> 	 integration model for  S.1 :
#> 		 integrate_parametric(formula = S.1 ~ BIP )
#> 
#> Risk models: 
#> 	 risk_binary( )
#> 
#> 	No estimates present, see ?ps_estimate.
#> 	No bootstraps present, see ?ps_bootstrap.
test + risk_binary() # same as above
#> Augmented data frame:  100  obs. by  6  variables. 
#>   Z Y    S.1    S.0 cdfweights      BIP
#> 1 0 0     NA -0.542          1 -0.51164
#> 2 1 0  1.095     NA          1 -0.00197
#> 3 0 0     NA  1.751          1  1.78313
#> 4 0 1     NA  1.062          1  1.03138
#> 5 0 0     NA -0.501          1 -0.58070
#> 6 1 0 -0.422     NA          1 -1.59308
#> 
#> Empirical TE:  0.744 
#> 
#> Mapped variables: 
#> 	 Z  ->  Z 
#> 	 Y  ->  Y.obs 
#> 	 S  ->  S.obs 
#> 	 BIP  ->  BIP 
#> 
#> Integration models: 
#> 	 integration model for  S.1 :
#> 		 integrate_parametric(formula = S.1 ~ BIP )
#> 
#> Risk models: 
#> 	 risk_binary( )
#> 
#> 	No estimates present, see ?ps_estimate.
#> 	No bootstraps present, see ?ps_bootstrap.