Calculate the Standardized total gain

Computes the standardized total gain for the risk difference. Optionally produces bootstrap standard errors and confidence intervals. The standardized total gain is the area between the risk difference curve and the horizontal line at the marginal risk difference. If the outcome is time to event then the STG is time-dependent, and a time point for evaluation is needed. If one is not provided then the restricted mean survival is estimated from the data and used.

Usage

calc_STG(
  psdesign,
  t,
  sig.level = 0.05,
  n.samps = 5000,
  bootstraps = TRUE,
  permute = TRUE,
  permute.times = 2000,
  progress.bar = TRUE
)

Arguments

psdesign

A psdesign object. It must contain a risk model, an integration model, and estimated parameters. Bootstrapped parameters are optional

t

For time to event outcomes, a fixed time t may be provided to compute the cumulative distribution function. If not, the restricted mean survival time is used. Omit for binary outcomes.

sig.level

Significance level for bootstrap confidence intervals

n.samps

The number of samples to take over the range of S.1 at which the VE is calculated

bootstraps

If true, and bootstrapped estimates are present, will calculate bootstrap standard errors and confidence interval.

permute

Not used, included for backwards compatibility

permute.times

Not used, included for backwards compatibility

progress.bar

Not used, included for backwards compatibility