Parametric integration model for the missing S(1)

Parametric integration model for the missing S(1)

Usage

integrate_parametric(formula, family = gaussian, ...)

Arguments

formula

Formula specifying the integration model for the surrogate under treatment. Generally the candidate surrogate will be on the left side in the formula, and the BIP or BIPs will be on the right side

family

Assumed distribution for the integration model. Must be compatible with the family argument of glm. Currently only Gaussian models are supported

...

Arguments passed to glm