Plot summary statistics for a psdesign object

Plot the treatment efficacy or another contrast of risk versus S.1 for an estimated psdesign object

Usage

# S3 method for class 'psdesign'
plot(
  x,
  t,
  contrast = "TE",
  sig.level = 0.05,
  CI.type = "band",
  n.samps = 500,
  xlab = "S.1",
  ylab = contrast,
  col = 1,
  lty = 1,
  lwd = 1,
  ...
)

Arguments

x

A psdesign object that contains a risk model, integration model, and valid estimates

t

For time to event outcomes, a fixed time t may be provided to compute the cumulative distribution function. If not, the restricted mean survival time is used. Omit for binary outcomes.

contrast

Name of contrast function to plot. "TE" or "VE" for treatment (vaccine) efficacy = 1 - risk_1(s)/risk_0(s), "RR" for relative risk = risk_1(s)/risk_0(s), "logRR" for log of the relative risk, "risk" for the risk in each treatment arm, and "RD" for the risk difference = risk_1(s) - risk_0(s). You can also pass a custom function directly as long as it takes 2 vectors as input (risk0 and risk1) and returns 1 vector of the same length.

sig.level

Significance level used for confidence bands on the contrast curve. This is only used if bootstrapped estimates are available.

CI.type

Character string, "pointwise" for pointwise confidence intervals, and "band" for simultaneous confidence band.

n.samps

Number of samples to use over the range of S.1 for plotting the curve

xlab

X-axis label

ylab

Y-axis label

col

Vector of integers specifying colors for each curve.

lty

Vector of integers specifying linetypes for each curve.

lwd

Vector of numeric values for line widths.

...

Other arguments passed to plot